requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
EbookLib==0.18
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
GoogleBard==1.4.0
googlebardapi==0.1.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
lxml==4.9.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.38
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-pptx==0.6.21
python3-openid==3.2.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
safetensors==0.3.1
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sqlparse==0.4.4
tokenizers==0.13.3
tqdm==4.65.0
typing_extensions==4.7.0
urllib3==2.0.3
wcwidth==0.2.6
XlsxWriter==3.1.2
yarl==1.9.2
